Fix hls support check
There was a case where the hlsSource was set and hls was supported
but the inferMimeType was ran and returned no support. This is because
the current inferMimeType only checks for the basename extension which
is a weak form of checking and a fallback of sorts. Do not try to
inferMime when the type is passed and supported.